Add components Board and Body
