+++ src/body/board/styles.module.css
@@ -1,0 +1,122 @@
+.bodyBoard {
+  display: flex;
+  width: 100%;
+  padding: 0px 16px;
+  flex-direction: column;
+  align-items: center;
+  box-sizing: border-box;
+  background: #fecaca;
+}
+
+.containerBoard {
+  display: flex;
+  height: 928px;
+  width: 1280px;
+  padding: 16px 0px;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 32px;
+}
+
+.containerTitle {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+}
+
+.title {
+  color: #171717;
+  font-family: Inter;
+  font-size: 24px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 32px;
+}
+
+.containerMenuCards {
+  display: flex;
+  align-items: flex-start;
+  gap: 32px;
+}
+
+.containerTasks {
+  display: flex;
+  width: 280px;
+  padding: 8px;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 8px;
+
+  border-radius: 6px;
+  background: #f5f5f5;
+  box-shadow: 0px 1px 2px 0px rgba(0, 0, 0, 0.25);
+}
+
+.containerFlex {
+  display: flex;
+  padding: 0px 8px;
+  align-items: center;
+  gap: 4px;
+  align-self: stretch;
+  justify-content: space-between;
+}
+
+.inputDisplay {
+  display: flex;
+  height: 35px;
+  padding: 8px 12px;
+  align-items: flex-start;
+  gap: 4px;
+  align-self: stretch;
+  border-radius: 6px;
+  border: 1px solid #d4d4d4;
+  background: #fff;
+  box-sizing: border-box;
+}
+
+.subTitle {
+  color: #171717;
+  font-family: Inter;
+  font-size: 18px;
+  font-style: normal;
+  font-weight: 600;
+  line-height: 32px;
+}
+
+.containerTask {
+  display: flex;
+  padding: 8px;
+  align-items: flex-start;
+  gap: 4px;
+  align-self: stretch;
+  border-radius: 6px;
+  background: #fff;
+  box-shadow: 0px 1px 2px 0px rgba(0, 0, 0, 0.25);
+  justify-content: space-between;
+}
+
+.task {
+  color: var(--foreground, #171717);
+  font-family: Inter;
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 28px;
+}
+
+.containerAddCard {
+  display: flex;
+  padding: 4px var(--spacing-2, 8px);
+  align-items: flex-start;
+  gap: 4px;
+  align-self: stretch;
+}
+
+.textAddCard {
+  color: var(--muted-foreground, #525252);
+  font-family: Inter;
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 600;
+  line-height: 24px;
+}
