+++ src/body/my-boards/styles.module.css
@@ -1,91 +1,3 @@
-.container {
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-  align-items: center;
-}
-
-.header {
-  display: flex;
-  height: 96px;
-  width: 1280px;
-  padding: 27px 0px;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.logoContainer {
-  display: flex;
-  gap: 16px;
-  align-items: center;
-}
-
-.logo {
-  width: 36px;
-  height: 36px;
-  flex-shrink: 0;
-}
-
-.title {
-  color: #171717;
-  font-family: Inter;
-  font-size: 24px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: 32px;
-  letter-spacing: -0.6px;
-}
-
-.botonContainer {
-  display: flex;
-  gap: 16px;
-  align-items: center;
-}
-
-.bAccount {
-  display: inline-flex;
-  height: 36px;
-  padding: 8px 12px;
-  justify-content: center;
-  align-items: center;
-  gap: 8px;
-  flex-shrink: 0;
-
-  border-radius: 6px;
-  border: 1px solid #d4d4d4;
-  background: #fff;
-
-  color: #262626;
-  font-family: Inter;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 20px;
-}
-
-.bLogout {
-  display: inline-flex;
-  height: 36px;
-  padding: 8px 12px;
-  justify-content: center;
-  align-items: center;
-  gap: 8px;
-  flex-shrink: 0;
-
-  border-radius: 6px;
-  border: 1px transparent;
-  background: #e5e5e5;
-
-  color: #262626;
-
-  font-family: Inter;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 20px;
-}
-
 .bodyBoards {
   display: flex;
   width: 1024px;
@@ -204,44 +116,6 @@
   align-items: center;
   flex: 1 0 0;
   align-self: stretch;
-}
-
-.containerMenuColor {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.colorMessage {
-  color: #262626;
-  font-family: Inter;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 14px;
-}
-
-.menuColors {
-  border: 1px solid;
-  border-radius: 15px;
-  width: 24px;
-  height: 24px;
-  position: relative;
-}
-
-.containerColors {
-  display: grid;
-  gap: 4px;
-  position: absolute;
-  top: 27px;
-  grid-template-columns: 1fr 1fr 1fr 1fr;
-  padding: 8px;
-  right: -52px;
-  box-sizing: border-box;
-
-  border-radius: var(--rounded-md, 6px);
-  background: var(--background, #fff);
-  box-shadow: 0px 1px 2px 0px rgba(0, 0, 0, 0.25);
 }
 
 .Fbutton {
