+++ src/body/styles.module.css
@@ -1,0 +1,7 @@
+.container {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  align-items: center;
+}
